+++ src/GitHub_Updater/Plugin.php
@@ -429,7 +429,7 @@ class Plugin extends Base {
 
 				// Unset if extended naming and same slug on dot org.
 				if ( isset( $transient->response[ $plugin->slug ]->id ) &&
-				     $this->is_extended_naming()
+				     ( $this->is_extended_naming() || $this->is_override_dot_org() )
 				) {
 					unset( $transient->response[ $plugin->slug ] );
 				}